plug the memory leaks
1) Keeping the MCJIT engine alive after compiling code is to expensive!
We need to delete it. But since it owns all generated code that would crash and burn. To solve this problem (for now) we override the destructor of our JITMemoryManager to leak the allocated code sections which allows us to safely delete the engine.

2) Calling selectTarget on the engine builder is not necessary and
additionally leaks the TargetMachine which keeps all relocation info
used during compilation alive.