remove builder from the passes and fix codegen
We cannot use the builder in the handler passes to get to the constant
pools. The reason is that the cp are stored in the Builder::Context
which are deleted after functionClose.

Instead we keep now a JITModule which extends llvm::Module to also keep
track of the relation between llvm::Function and nativesxp.

This commit also fixes many bugs which where introduced by the botched
up refactoring two commits ago