Cleanup switch to instantiated Patterns
We decided to actually instantiate the ir::Patern and not just have
shallow wrappers around llvm::Instruction.

But the change was only half way pushed through. Especially the ::create
functions still returning value types and references causes a lot of
confusion. And everything outside ir::Intrinsic was not added to the
rjit ir at all.

This patch polishes those parts. Also it adds a new MultiPattern which
is an rjit Pattern for multiple llvm::Instructions. The design for
MultiPattern is not to be seen as final, but merely a first iteration to
make it work with the multi-llvm-instruction patterns i want to
introduce.