fix no-segfault test
the gnur testharness does fuzz all public apis of packages. thus we need
to be a bit careful to not assert in our api.

also we have to make sure the test harness does not fiddle with
enable-/disableJIT(). to fix this there is now a sticky enable jit
version.