Complex assignments: we have to rewrite the target too
In the case of specials the target gets reevaluated. Eg. in the
case of:

a[1] <- 2

which translates to

[<-(a, 1, 2)

we would evaluate a, then passing this ast to the special, which
(not knowing about our evaluated a) evaluates it again...

therefore we need to do the same thing we do for the value, ie. craft a
synthetic ast, with the value already in there.

The ast interpreter uses a different strategy: it replaces the first
argument by *tmp* and binds the intermediate result to this variable.

We'll see if we can do without it...