fix merging of environments
if we merge two environments and in one a value is missing and in the
other it is present, then it is incorrect to just copy over the
non-missing one. Consider the following code:

if (x)
  a <- 1
a

Such a strategy would declare a to be constant one on line 3. Therefore
we require AVALUE used in State to have an 'Absent' value. The merging
will merge the value defined in just one branch with AVALUE::Absent()
instead.