bugfixes and code for local variables
this code introduces (currently still disabled) code for discovering
local variables and runtime guards to detect if they get modified from
the outside.

also we change the return value of analysis[symbol], if there is no
information for the symbol. It used to be bottom, which is confusing and
in most situation incorrect. For example a dataflow analysis of:

a <- a || 2

would determine a is "bottom (meet) const" -> "const" which is obviously
incorrect. therefore it now returns top.