assignment rhs has to be marked as shared
the rhs of an assignment instruction has to be marked as shared.

the reason is that the result of an assignment is the rhs. For
example consider:

    a <- b <- c

this bug was uncovered by the previous change.

additionally this commit removes the bogus optimization for
removing uniq's and introduces a better optimization where
dup/pops are removed if the only use of the value is
set_shared or uniq instruction (ie. the copy is never used).