cache bindings of the local frame
a very small cache per eval invocation keeps track of the most recently
looked up variable from the inner most environment. The location of
the binding in the environmen is stored in the cache. This saves us
looking through the linked list of bindings in the local env.

The mechanism is very similar to the gnur BC interpreter. Though a bit
more lightweight.

long term something more intelligent would be nice. But for now this
seems like a good addition. Speeds up some things by 20%.