Add dispatch tables to closures.
This is part of a larger body of work, but is being merged now at an
intermediate point.

Before, the body of a closure would be a function. Now, the body of a
closure is a dispatch table, which contains different versions of a
function. A version may be based on optimizations, or a different
calling convention for that function (e.g. one that does not allocate
promises). Note that a function has a reference to its closure, so it is
always possible to go from a function to its dispatch table.

The GC has already been updated so it can trace through the dispatch
table:

    https://github.com/reactorlabs/gnur/commit/b539953

When a function is called, the first entry of the dispatch table is
selected. Currently, there is no mechanism for selecting a different
version from the dispatch table.

The optimizer is still called as before, so a new function version is
compiled and inserted into the table. To grow a dispatch table, a new
one is allocated (with twice the original capacity) and entries are
copied over.

Currently, the optimized version of a function is never called. This is
to avoid making additional changes to the deoptimizer (which would be
reverted later in this work).

There are some "leftovers" remaining in the code. These have been noted
and will be addressed later, since the necessary changes are quite
invasive.

Roughly, the next steps of this work are:

  1. Generate different versions of a function, to insert into the
     dispatch table. Describe the version's capabilities (e.g.
     optimization level or calling convention) as some "identifier"
     field.

  2. Add a mechanism for calling a particular function version. At
     first, this mechanism will be slow, as it scans the entire dispatch
     table. Also, deoptimization needs to be updated.

  3. Improve performance, e.g. with an inline cache.