Move the formals handling to interpreter.
The compiler now compiles the formals (if present) and just stores the Code objects in the Function. It does not change the formal argument list of the compiled closure.
Only in interpreter, when the missing arguments with default values are needed, we pick the Code objects and wrap them in promises.