Expose SEXPs in RIR objects to R's GC.
This patch adds a header to all RIR objects that contains information for the GC about pointers that we want to trace.

Every RIR object (with the exception of Code, since it is not a valid SEXP), now has to have rir_header as its first field and fill it with
1) offset in bytes to the first SEXP to trace
2) number of SEXPs to trace

Then it should make sure that all traced SEXPs are in one continuous block of memory.

For now, this applies to DispatchTable and Function, and to function Signature as soon as we have it.