Move magic number to header
This patch removes magic numbers from DispatchTables and Functions and moves it the header (info field).

The reason is that up until now, when we checked if something is a particular type of object, we looked at different offsets (the magic numbers did not overlap).
This caused a rare bug when a Function got garbage collected and a DispatchTable then allocated in its spot. If we now checked the pointer to see if it was a Function, we saw the old memory still containing the correct Function magic number (the DispatchTable is shorter and thus did not overwrite the magic). This lead to memory corruption.

Now the magic is placed in the object header and so is always overwritten correctly.