let's clean up the horrible naming around `[` and `[[`
internally we called `[[` extract and `[` subset, but then
`[[<-` and `[<- both subassign. In the formar case we used
1,2 as a postfix to denote 1 vs. 2-dimensional indexing,
whereas in the later case we used it do distinguish between
`[` and `[[` :/

this commit cleans is up. Now both `[[` and `[` are called
extract. The former is extract2, the later is extract1.

Then we use a second index for the number of indices. So
`a[1,2]` is extract_1_2 `a[1]` is extract1_1 and `a[[1]]`
is extract2_1.