unify the implementation of all BB Visitors
There are 3 parameters for BB visitors:

1. performing an action (ie. unconditionally run on all BB's) or a
   predicate (ie. run until the first action fails).
2. rely on BB ids not to change during the visit
3. stable or random visit order

This commit abstracts the visit algorithm and makes those 3 things
template parameters.

(1) is reflected in the 'run' vs. 'check' api and implemented
by a different 'ActionKind' parameter to 'runGeneric'.

(2) is realized by the 'Marker' parameter and the two implementations
'IDMarker' and 'PointerMarker'. Finally, (3) is given by the
'STABLE' bool.

To avoid clients having to deal with the template parameters, there
are predefined versions 'Visitor', 'BreadthFirstVisitor', and
'UnstableIDsVisitor'.