Closure conversion, sort of ;)
When we know that a load will hit the outer environment (called
theContext, or ? in pir), we do not want to keep the local
environment alive. For example in `function() a`, we know that
`a` is gonna be loaded from the closure environment.

So this optimization transforms:

    env     e0.0  = MkEnv            (parent=?)
    val^?   %0.1  = LdVar            (a, e0.0)

to

    val^?   %0.0  = LdVar            (a, ?)

track "static" environments