Add support for static calls
When the call target is statically known, let's use a different call
instruction that allows us to directly link to that known closure.

This commit also fixes a bug in the compiler, where we assumed that
every 'static_call_stack' calls a builting (which is wrong).

Also we add support for speculative static calls for callsites which
have been observed to be monomorphic.