A new Register allocator (#120)
* A new Register allocator

This patch adds a new liveness interval algorithm, a naive algo for
sinking ssa variables to stack slots and a coloring register alloc
for the remaining variables.

Instruction -> BB -> {live, start, end}

Two variables do not interfere if: for all BB: either they are not
both live, or their start-end intervals do not overlap.

Stack slots are assigned for variables which are produced and
consumed exactly once in the same BB. Additionally trivial phis,
which are produced at the end of all input BBs and the consumed
by a phi at the beginning of the successor block are sunk to stack
slots.

The approach is fairly naive and should be improved.

The remaining registers are allocated in a linear pass on the
dominator tree. According to Hack and Goos, this should be optimal.

Caveat: we precolor all phi, such that phi and all inputs are
coalesced into the same input. Also as of yet, we do not convert
to CSSA. I conjecture that there should be a program that breaks
this approach. Something like:

    x1       x2         x3

      \      /         /
         y = x1+x2
            \        /
           x4 = phi(x1,x2,x3)

But so far I haven't been able to construct it, or understand
why it couldn't happen. More investigation is necessary and
probably another reading of Boissinot etal.

Unfortunately some other things got lumped into this PR:

* The verifier was not run after every optimization
* Delay instruction did delay phis (which is never a good idea)
* Scope resolution was broken for `LdFun`. This instruction
  performs an implicit `ChkClosure` test. If we replace a `LdFun`
  with something that is not of static type `closure`, we need
  to add that runtime check.
* Scope resolution stupidly places phis. Changed it to move
  phi's up the cfg to the point where the inputs actually come
  from different predecessors.