refactor call bc implementation and support full pir CC
this heavily refactors all the machinery behind the call bytecodes.
The main concept is, that all the arguments related to a call site are
put in a stack allocated CallsiteContext struct, to avoid super-long
argument lists. Also we use the type of those structs to statically
advertise, if we have yet to search for a callee (DispatchContext)
or not, or if we already popped eager args off the stack (the Final... variants).

As a new feature this version has the dispatch method, that selects
slot 1, only if possible. Also we now support eager arguments for slot 1.

It is implemented the following way: rirEvalCode takes SEXP env and
CallsiteContext* callCtx, either of them can be null.

* If env is null, the callee must set env before using any envy bytecode,
  like stvar. Pir optimized code does so.
* If callCtx is null, the callee must not use ldarg_. Slot 0 code does so.
* When ldarg_ accesses an arg, it will query callCtx->hasEagerArgs and
  either create a promise on demand, or load the value with
  callCtx->eagerArg(i), which reads it from the stack. The caller
  must pop the arguments after the function returns.