Gracefully fail pir compilation on unsupported features
Refactor the rir2pir compiler such that compilation can fail. In which
case the compilation is aborted in a graceful maner.

This patch compiles every rir compiled code also to pir, to catch
errors in the compiler. The compilation is allowed to abort, but should
not crash.

Compilation back from pir to rir is not yet enabled.
Optimizations are, though.