towards enabling pir2rir
For every piece of code we compile, assert that it can be successfully
compiled to pir and back to rir. The result is thrown away.

For this change to work a couple of bugs are fixed:

promises can be recursive. We can refer to a promise in a promise.
Therefore it is not possible to compile promises in a fixed order.
We need to compile them on demand.

(Recursive dependencies should not exist)

Support for environment arguments was not very robust. This patch
simplifies the code by abstracting the handling of arguments somewhat.

Before every pir instruction, we always load all arguments; after every
pir instruction we always store the result.

The special cases are:

* Some instructions want the argument implicitly. If an instruction
  takes an environment and does not consume it explicitly, then
  there is a preamble that loads the env and issues a setEnv.
* MkArg is currently a bit of a hack, and it still is in this patch:
  either, we have an eager arg, then the result of mkarg is that
  eager arg. Otherwise it does not have a result.
  Then the call is responsible for consuming all the eager args.

The implicit environment loading gets very annoying, if the environment
is not the TOS. Therefore this patch changes the order of vararg
instructions, such that (as in the fixed argument length case), the
environment always comes last.

This allows us to evaluate the arguments in the pir-order, when
translating back to rir; and we always find the env to be tos,
therefore to consume it, we just issue a set_env_

Our guard function was using `==` to compare closure values. That
fails for primitives. Add a new `identical_` bytecode that does
pointer equality.

Currently we don't have deoptimization supprt. We'll need to add
it soon. This patch just aborts instead of deopting.