proper call instructions for pir
For too long we tried to workaround the semantic mismatch between
lazy pir and rir call instructions:

rir call instructions have arguments implicit, pir call instructions
have them explicit, with MkArg that creates a promise.

This commit adds the missing instructions to rir, such that we can
explicitly create promises and pass them as arguments.

In a later step, we could either completely remove `call_` from
rir, or we could also have a special pir `OnlyLazyArgsCall`
instruction in pir, that we can generate in a later pass and
then compile to the `call_`.