most binops need an environment
In general binops need an environment. Only when we know the type of
the lhs, we might be able to get rid of it.

This change just makes it sound for now. In the future the idea would
be to have a pass that replaces the environment with nil, when we
can infer from the types, that it is not needed.