Add Env variable to enable PIR optimizations
The env variable is called `PIR_ENABLE`

There are 3 settings:

on           : compile on second invocation
force        : always compile with pir
force_dryrun : always compile with pir, but never use the result

'on' and 'force' do not pass the gnur tests yet, that's why the default is off. All options however pass `bin/tests`.