Rewrite source maps and remove nops in the function writer
1. Drastically simplifies the source maps. Instead of skiplist and
linear source maps, we just keep a sorted list of pc to src idx
mappings. When looking up a source we perform a binary search
through this list.

2. The recent removal of BC cleanup left a lot of nop's in the
BC stream. This patch removes them in the function writer, when
we serialize a BC stream. The whole thing turned out to be tricky
since there are offsets that need to be updated on the fly.
First of all the sources are relative to the old BC, but more
annoyingly, jump labels and patchpoints offsets need to be
recomputed on the fly.