Cleanup in preparation for typefeedback (#177)
A Bunch of cleanups, in preparation for collecting typefeedback.

I wanted to simplify the Code struct. For that I removed the CallSite struct part. This means that I put the main parts of the CallSite struct into the BC stream, and the non-essential parts I threw away for now.

I think this makes a lot more sense, but this also means, that i had to add 2 BC and 3 BCs became variable length. On the plus side, the BC objects from ir/BC_inc.h now really capture the whole BC, with all arguments, and it is no longer necessary to pass around CallSite pointers with it.

Also, I removed the CodeEditor, since that one was quite intertwined with the CallSite structs.

The idea is to later add a new (and better designed) part in the Code object, where we can store additional metadata for any BC. But the idea is, that the metadata is not necessary to interpret the behavior of a BC!