Add pir support for deoptimization
Add a deopt pir instruction. Currently this one does nothing more
than changing the pc and code pointer. All the smarts are in pir,
where we make sure that an env is created and all the values on
the stack are captured.

Currently we do not inline the deopt instruction. Thus we do not
need to synthesize call frames. I plan to support those features
in a subsequent patch.