Introduce RirRuntimeObject superclass
RirRuntimeObject is the superclass for RIR objects embedded in
EXTERNALSXP, e.g. Code, DispatchTable, and Function. The constructor is
protected, so it can only be instantiated by its subclasses.

All the common code has been moved to RirRuntimeObject, e.g. container,
unpack, and check.

The constructors for RIR objects are now a bit more complicated. Error
reporting will be slightly worse (now it's not clear which type of RIR
runtime object had an error), but a stack trace should provide the
missing information.