Remove cppcheck pre-push-hook
Turns out cppcheck is brittle across different versions. 1.84 uses a
different suppression syntax than the version installed on Travis, and
reports a different set of warnings.

Keeping everybody's cppcheck versions and codebase in sync is not worth
the effort. The One True Cppcheck will be the one that runs on Travis.