Named arguments for PIR Calls
Introduce a new NamedCall PIR instruction that supports named
arguments. Currently this named call instruction is a completely
separate instruction. Therefore for existing optimizations and
analysis, this new instruction is completely opaque.

What we gain from this commit is that named calls can be compiled
to pir and back to rir.