Check exit code of all R package tests and deal with the fallout
The gnur testrunner only returns non-zero result when running with TEST_MC_CORES > 1. That's why the package tests failed in the past with TEST_MC_CORES > 1, since they where actually always failing :)

To be sure we always fail when we should (even when the cores are set to 1) this PR adds an additional script that greps the test logs for ERROR. This script has the additional advantage that it prints the full test error in the log, where it is not mixed with output from other concurrently running tests.

A number of fixes where neccessary after "enabling" those tests for real:

* install all the latex crap on travis
* start a virtual framebuffer for the tests
* creating a R context before calling dispatch in the subassign bytecodes
* preventing the closures mentioned in the typefeedback from being collected. For this i added a generic "extra pool" to the code objects. This is a list that you can add stuff that needs to have the same lifetime as the code object.
* #234
* fixed handling of missing arguments passed to ldarg
* standardGeneric cannot be inlined (like usemethod)

Since now all the issues are resolved, this PR scales TEST_MC_CORES to the number of available CPUs, to make the tests run faster. Overall it unfortunately now takes much longer, since installing all the latex crap and running the previously failing latex tests takes so much time.

This PR also makes the promise in the MkArg instruction private. This turned out to be a red herring, and there was nothing wrong with the updating of this field. I still left the changes in the PR, since it might prevent future errors.

I also removed a redundant entry in the travis build matrix. the tests are taking already long enough...