add framestate to the call instructions
for inlining we need a framestate that corresponds to the call instruction.
we used to just put a framestate after the call and then let the inliner
linearly search for it. This commit places the framestate right before
the call (which is more correct) and adds it as an argument to the call
instruction. thanks to that it is always unambiguous which framestate belongs
to the call.