inlining promised calls
there was an issue with inlining functions such as `a` in:

    a <- function() 42
    f <- function(x) x()
    f(a)

after inlining `f`. The reason was that after inlining `f` we get
`ldfun x`. Even if we manage to trace this through the promise
`ldvar a`, we do not inline the promise, because there is no force
instruction. The reason is that `ldfun` itself forces, because of
the strange 'skip-non-closure-bindings'. Also, due to those semantics
we can only replace `ldfun` if we are sure that the thing we load
is a closure, which we can only after inlining the promise. A
chicken-egg problem.

This is resolved in this commit, by adding a "guessed binding" input
to the `ldfun` instruction, where we can add the most inner binding
that matches the ldfun, try to infer its type over time and finally
if we succeed to prove that it has type closure, replace the `ldfun`
with it (or discard it if not).

Additionally this commit removes unused checkpoints (such that they
do not block optimizations) and marks framestates as non-effectful (such
that they get removed automatically).

(also adds some tests for this and constantfolding)