improve analysis precision a lot
the main change is that during the analysis we keep track of the
return values of interprocedural and inter-promise analysis and
feed the approximated result value back to the caller.

This allows us for example to de-promise promised ldvars and also
supports interprocedural type inference.

as an additional trick we keep track if a promise might have used
reflection. if a promise could not have used reflection, then we
do not need to create a callee env for it.