more precise drillDown method for scope analysis
in scope analysis we use the drillDown helper to find out
if we somehow already computed a value. For example if
we want to interprocedurally follow a call, but the calle is
just a `ldFun` instruction, we use drillDown to find out
if the current analysis state permits looking up the function.

So far we only did this when drillDown finds an unambiguous
single value. But for some things (like finding out if a
promise executes or not) it is enough to get the AbstractPirValue.

Therefore this commit extends drillDown to also return abstract
results.