slightly more fine grained named controls
we have the setShared instruction which bumps the named counter.
but there are several instances where we just want to ensure that
it's named, ie. >= 1. So setShared is very conservative...

this commit adds the ensureNamed instruction which, well, ensures
that named is >= 1. We also fix a missing ensureNamed in the for
loop.

Thanks to this fix, the named counts are now "more correct"
and I thus managed to avoid some allocations in the interpreter.
We can now trust the NO_REFERENCES macro in the interpreter and
just go ahead and freely reuse objects with that property.