cleanup static analysis
Static analysis keeps snapshots of the state at all merge points.
Additionally also keeps interprocedural analysis state.

This patch uses a struct to make it a bit more readable, where we
store the initial state of a BB and where we keep the extra
snapshots after interprocedural calls.

Also let the analysis implementation decide when to take a snapshot
(instead of after every call, like we used to).