dynamic dispatch based on signature
in this commit we order versions in the dispatch table by increasing
number of assumptions. To dispatch we search backwards through the
list. Entry 0, the baseline, always matches.

The assumptions are just a static bitset. Probably we want something
more elaborate, but since we are not exactly sure what we want, it's
probably a good first iteration.