use a ldvar in guards
instead of using ldfun, we use ldvar to load the function
we guard. that is conservative (it might fail more than
neccessary, if there is a non-closure binding in local scope)
but otoh it has no side-effects, thus should give us more
abilities to optimize.

Also in this pass, start a list of safe builtins and create
a factory to decide whether a safe or unsafe builtin call
has to be inserted.