fix soundness issue in static analysis
currently only return statements are merged into the exitstate in
the static analysis. while this might be correct for some analysis
it is certainly not correct in general.

for example the dead store optimization looks at result() of the
analysis to determine whether a load is visible. If the load would
be in the deopt branch we would not see it.

I looked at all the places where we currently use result() and it
seems to me, that including the deopt state is the correct choice.

thus let's always include the deopt state for now and we can still
make it more clever in the future.

Also, this revealed a case in the scope analysis, where we did not
consider that a deopt influences the return value of a function.