builtins are eager
if we speculate on builtins, we should eagerly evaluate all args.
This was missing before, so we would actually pass unevaluated
promises.

To be able to do this, we have to put the guard earlier, ie. before
the arguments. because if it turns out not to be a builtin, forcing
early would be wrong.