make pir typesystem differ between values and promise wrapped values
A R value can be:

1. an unevaluated promise
2. an evaluated promise
3. a value

In the past we did not differ between 2 and 3 which can lead to
confusing situations and easily leads to a situation where we
put the wrong kind of value in the wrong place.

This commit adds a new type flag `promiseWrapped`.

Therefore:

1. maybeLazy && maybePromiseWrapped
2. !maybeLazy && maybePromiseWrapped
3. !maybeLazy && !maybePromiseWrapped

To transition 2 => 3 a Force instruction is always needed, because
many internal instruction expect either one or the other.

From 1 => 2 it is enough to prove that a value is already evaluated
and then remove the maybeLazy flag.