support for missing arguments in pir with assumptions
In the cases where we are absolutely sure which arguments are missing
we can easily support them in pir, by explicitly creating them.

To know which args are missing, we need to know that there are not
explicitly missing args (ie. no f(,,)) and we need to know how
many arguments to few where supplied.

Maybe we should make missing args a bitset? then we could drop the
first assumption.

[wip] add some preliminary support for missing_

fix issue with ldvar as default args

fix visibility

finally we see issues with visible in the testsuite

maybe getting there

another try

fix inlining of nargs

quickly patch the worst recompilation pathologies