it's hard to guarantee that static dispatch never fails
The following can happen:

we have some available assumptions, compile a call based on those
assumptions. But then a later transformations changes the code in
such a way that we fail to statically show that the assumption
holds.

This case is really hard to categorically rule out...

so let's just deal with the fact that once in a blue moon it
might happen that we cannot dispatch.