Late speculation on builtins, in eagerCall pass
If we have some builtin call, but we can't statically prove that the call
does go to a builtin, then we would like to speculate. Because speculating
on a builtin allows us to eagerly evaluate all arguments.
So far we had to decide on whether to make this speculation in the first
pass of the compiler. This PR presents the first case where we are able
to do this speculation late in the pipeline. This helps us if early we
are not able to speculate due to a missing checkpoint. For example if we
are compiling promise code.

To be able to support this case, we need to find out if there is a
deoptimization point available. The deoptimization point can come from
anywhere before the speculation, but after the last effect. It is always
better to use an earlier deoptimization point, because that leads to less
deoptimization points being used, thus more that can be removed, thus less
restrictions for the optimizer. I added a simple `AvailableCheckpoints`
analysis, that answers that question. The api is
`AvailableCheckpoint checkpoint(code); Checkpoint* res = checkpoint.at(instruction); if (res) ...`.

In the same PR is also improve the StaticAnalysis framework a bit,
to speed up random access queries, like in this one. (For the existing
optimizations it was enough to be able to quickly iterate over the analysis
result at every instruction in sequence.)