Increase max compilation unit size and fix a number of bugs in the compiler
Due to large methods being excluded from compilation a number of
bugs managed to sneak in. They got revealed when I tried to
increase the limits. This commit fixes:

* the RIR push_ bytecode changed the visibility flag. However pir
LdConst of course is effectless. This can lead to missed invisible
settings. This is solved by making the visible explicit in rir already.
I fixed this, because it spams the test logs with, with NULL's from
dangling else branches in `cleanEx`, which lost their invisible.
* The above fix kills some optimizations. Thus i improved the visibility
analysis a bit. But also, We don't really care about
visible that much. We just want to get it right most of the times. Thus
this commit ignores visibility in some strategic cases. For example
promises which are pure except for visibility change, we still inline
(the subsequent call will change visibility anyway). Checkpoints also
ignore visibility. Worst case we deopt with the wrong visibility
setting.
* The scope analysis would sometimes generate pretty terrible phis, that
caused the lowering in the allocator to just diverge. Excluded those
kind of phis in the verifier and fixed the scope analysis to not produce
them.
* cleanup would happily merge the block with a phi input and the block
with the actual phi. this leads to input+phi ending up in the same
block.
* banned a couple more reflective functions from being inlined