add contexts for inlined functions
If we inline functions, and those functions are not leaf functions (ie.
they have a call instruction in them), it is not sound to not allocate a
context (ie. a RCNTXT) for them.

If for example a callee of the inlinee does reflection on the inlinee,
it will see the context of the caller instead. Something like:

    Caller    -->     Inlinee   -->   CalleeOfInlinee
    [    just one context   ]   -->   [  sys.parent ]

In this case sys.parent returns the context of Caller instead of
Inlinee.

To solve this issue an explicit push_/pop_context instruction is added
to pir and rir. On the rir side we add an optimization to remove
contexts again, if after optimizations the inlinee does not have
outgoing calls anymore.

On the rir side we use an inlinedContextTrampoline (similar to the loop
context trampoline) that recursively calls the interpreter function,
while keeping the current interpreter state (like local variables, pc,
stack and so on). The corresponding pop_context terminates that
recursive invocation and continues execution in the original call frame.