fix for an impossible to debug problem
the problem was mk_env_ reordering, which caused the wrong environment
to be created first and thus resulted in a context with a wrong env.

For example in:

```
e1 = mkenv
     pushcontext
e2 = mkenv
```

we would reorder to:

```
     pushcontext
e1 = mkenv
e2 = mkenv
```

which causes the env e1 to be stored on the inner context and e2
nowhere.

To fix the issue I added a `context` number to mkenv. Thus we get:

```
e1 = mkenv         context 1
     pushcontext
e2 = mkenv         context 1
```

and after reordering

```
     pushcontext
e1 = mkenv         context 2
e2 = mkenv         context 1
```

as can be seen we update the context position, when hoisting over a
pushcontext (or when removing push/pop contexts). Thus the mkenv always
updates the correct environment.

This also solves the issue where a deoptimization recreates an
envrionemnt. For example we sometimes have:

```
  cp1 = checkpoint  DEOPT | CONT

DEOPT:
  e1' = mkenv     context 1
  deopt

CONT:
  e1 = mkenv      context 1
  assume(cp1, ...)
```

here we create the env `e1` twice in the deopt case. Since we are now
precise in tracking the contexts a mkenv belongs to, we can safely
override `e1` on the context with `e1'` in the deopt case.