Hopefully fix the allocator
There was a case where the allocator merged two inputs. This refactoring adds a new phi to prevent that.
The problem was in phi(a:bb1, b:bb2) when bb2 was not immediate predecessor of the phi block and b was only created conditionally.
Then in the case where b is not actually created, a should flow into the phi.
Now we add a new phi that distinguishes a and b in this situation.