Set default filters in DebugOptions constructor.
The single-argument constructor for DebugOptions initializes the flags,
but not the pass filter or function filter. This means any matches
against the filters will always fail. This is very surprising and
difficult to diagnose, so the default filter should be ".*" and match
everything.

Currently, nothing uses the single-argument constructor, but it has been
used in the past, and without this commit it's a bug waiting to be
discovered...