delay environments into deopt branches and GVN
this pr adds a neat trick to scope analysis: basically if an environment
is only needed for deoptimization (ie. only occurs in framestate and
stvar), we try to recompute it in the deopt branch.

this leads to redundancies. because we first elide ldvar, and then later
synthesize the environment with the same values. thus we create the
almost exact same phis twice. to solve this problem i added a very
simple gvn.

also contains a number of fixes to the phi placement and scope
resolution.