fix and improve static reference count
to compute the static reference count we need a static analysis. The
reason is that in

    create a
    loop {
      consume a
    }

we must know that a is used multiple times. Our old one-pass analysis
would think that it is used only once, even though it is in a loop.

on the other hand we can improve the precision with this commit. In both
cases:

    create a
    if (...)
       consume a
    else
       consume a

and

    loop {
      create a
      ...
      consume a
    }

the value a can have only one use at runtime.