inc is intended to increment in-place
I remembered, that inc_ is supposed to update in-place, even if the
refcount is 1.

That was a feature, not a bug, to be able to increment the loop counter
without creating a new vector all the time.

So the idea is we compile the for loop as follows:

    %i = 0
    while (%i < size(seq))
      ...
      %i++

now, we need to mark the (stack) value %i ensureNamed, because
otherwise the < comparison might override it (values with refcount 0
can be reused). But then the i++ at the end of the loop, should
still be able to increment i without duplicating it. so the inc
bytecode is allowed to destructively modify it

but, the pir version of inc was broken, because the pir optimizer did
not preserve this invariant, that an inc'd value is not used anymore
after the operation.

for now I replaced it by a normal add. this might have potential
performance issues, but we need to solve them differently.