re-add inc and fix refcount for subassign
subassign are allowed to override values with refcount 1. This is to
ensure that:

    ldvar i
    subassign ...
    stvar i

does not actually need to copy the vector, even though it has a name
(here "i"). But of course if pir would end up optimizing a subassign and
e.g. do:

    ldvar i
    subassign ...
    dup
    stvar i
    subassign ...
    stvar j

then this would be catastrophal. So, for this case we need to ensure
that the refcount is actually at least 2 for the subassigned array.

The same is true for the inc instruction.