fix named counting around stvar
stvar had a completely bogus ensure named, that cause it to always
duplicate the assignee.

This completely defeated the in-place update of subassign, which masked
3 bugs there (see new comments in the Compiler.cpp and interp.cpp).

To make the inplace update of subasign sound, a new ldvarForUpdate
bytecode ensures that a variable loaded from an
outer environment gets its named count increased.

improve static refcount analysis
also speed up merging and set an upper limit for the abstract state
size.