fix static refrence count
discovery of phi congruence classes was not working properly, since we
did not follow PirCopy instructions in all cases. This lead to the zizi
variable in mandelbrot not being ensureNamed. The first use of the zizi
variable is actually *before* the problematic version `zizi = zi * zi`.
This use was not counted as a use, because on the first use we did not
correctly follow the phis, to reach this declaration.